feat: Delete AzureOpenAIEndpoint when deleting a model
This commit adds functionality to delete the related `AzureOpenAIEndpoint` when deleting a model with the provider set to 'AzureOpenAI'. If an `AzureOpenAIEndpoint` is found for the model, it is removed. If no endpoint is found, a log message is printed. After deleting the endpoint, the main model is then deleted.