refactor: Use deleteOne() method to remove AzureOpenAIEndpoint
This commit refactors the code in the models.js file to use the deleteOne() method instead of the remove() method when removing the AzureOpenAIEndpoint. This change aligns with the MongoDB driver's recommended approach for deleting documents. By using deleteOne(), we ensure that only one document is deleted, which is the desired behavior in this case.